perf(core): getType uses a cache for well known types.
getType is called a lot, and it just run the same regexp over and over
on the same base types.
A cache increase its own efficiency by more than 80% for basic types.
On a real world application with a lot of components,
getType was profiled for 8% of call duration before this patch,
and about 0.5% after.
The impact is more limited for smaller applications.